executor: fix DATA RACE caused by change the global logger whe… (#16722)
